+++ src/declaration/index.d.ts
@@ -53,9 +53,17 @@ export function html(
 ): Promise<string>;
 
 /**
- * Registers name of css file to be loaded on next rendering.
+ * Lazingly load CSS to page.
+ * @param {string | Array<string>} css - name of css files that need to be
+ * lazy loaded.
+ * @returns {(immediately?: boolean) => void} function that unload registered css
+ * (removes from the page). If [immediately] is `true`, then css will be removed
+ * in time of function's invoking, otherwise css will be removed on next
+ * rendering step (`Router.to`, `Router.reload` etc).
  */
-export function registerCss(css: string | Array<string>): void;
+export function registerCss(
+  css: string | Array<string>
+): (immediately?: boolean) => void;
 
 export type Route = {
   path: string | RegExp;